+++ src/FeatureExtractor.h
@@ -4,7 +4,7 @@
 class FeatureExtractor{
 	public:
 		//FeatureExtractor(std::string name,Input* i);
-		virtual bool extract(int* r){return NULL;};
+		virtual bool extract(double* r){return NULL;};
 		std::string getName();
 	protected: 
 		std::string featurename;
